More progress on logging in.
Once I get the php files moved over it should work. Right now the main thing opening is the drawer. Still working on getting pieces of that to work. I took out the settings button on the toolbar and replaced it with a camera but it doesn't do anything yet.